The phase ONE of a sorting is complete... Basically it sort all the tuples that can fit in one the memory! In order to do so, the user is asked to given the number of buffer.
But if the whole unsorted result cannot fit in the memory we have a similar problem than before i.e. the sort is applied only for a group of tuples.

To solve the problem we should implement the phase TWO of the sorting algorithm i.e. making as many runs as necessary to sort the whole file. But to do it?

N.B. You can deduct than in the current version, if you want the whole file to be sorted you should take enough memory. i.e give a batch size and a number of buffer large enough!