PHASE TWO OF SORT HAS BEEN DONE....
after phase one I write the sorted tuples in a external file.
In phase two I merge them...

The algorithm sorts on several attribute level and DESC is not implemented yet...

BUT THE ALGO IS WRONG... ALL MY PAGES ARE OVERFLOWED... anyway... it sort ASC in any case so we can go on the other tasks. I think I need your help to solve this problem