PHASE ONE has been improve with a Comparator and now works with ASC and DESC.
PHASE TWO still works and takes as well ASC and DESC...

BUT I AM STILL LAZY... the writing of a file is done at once instead of doing it page by page when the outbatch is full.

Never the less we can go on on the project. I am dealing with this matter in parallel.